+++ ArWikiCats/legacy_bots/ma_bots2/general_resolver.py
@@ -15,7 +15,7 @@ from ...helps import logger
 from ...time_formats import time_to_arabic
 from ...translations import Jobs_new, jobs_mens_data
 from ...utils import get_relation_word, get_value_from_any_table
-from ..ma_bots2.ar_lab_bot import find_ar_label
+from .ar_lab_bot import find_ar_label
 from ..make_bots.bot_2018 import get_pop_All_18
 from ..matables_bots.bot import Films_O_TT, players_new_keys