+++ ArWikiCats/legacy_bots/ma_bots2/lab.py
@@ -19,9 +19,9 @@ from ...translations import (
     religious_entries,
 )
 from .. import tmp_bot
-from ..ma_bots.country_bot import fetch_country_term_label, get_country
+from .country_bot import fetch_country_term_label, get_country
 from ..make_bots.bot_2018 import get_pop_All_18
-from ..ma_bots import bys
+from . import bys
 
 
 @functools.lru_cache(maxsize=10000)