+++ ArWikiCats/legacy_bots/with_years_bot.py
@@ -14,7 +14,7 @@ from ..helps import logger
 from ..legacy_bots.common_resolver_chain import get_lab_for_country2
 from ..new_resolvers import all_new_resolvers
 from ..translations import WORD_AFTER_YEARS, get_from_pf_keys2
-from .ma_bots import general_resolver
+from .ma_bots2 import general_resolver
 from .make_bots.reg_lines import RE1_compile, RE2_compile, RE33_compile, re_sub_year
 from .matables_bots.data import Add_in_table
 from .political_terms import handle_political_terms